+++ src/index.js
@@ -6,6 +6,7 @@ import { constants } from './constants.js';
 import { Helper } from './pf2e-subsystem-helper.js';
 
 let subsystemTabActive = false;
+let appID
 
 Hooks.on('init', function() {
 	game.settings.register(constants.ID, "Reputation", {
@@ -55,7 +56,7 @@ Hooks.on('init', function() {
 Hooks.on('renderPartySheetPF2e', function(partySheet, html, data) {
     if (!game.user.isGM || !(game.settings.get(constants.ID, "Reputation")||game.settings.get(constants.ID, "Influence")||game.settings.get(constants.ID, "VictoryPoints")||game.settings.get(constants.ID, "Chases")||game.settings.get(constants.ID, "Infiltration")||game.settings.get(constants.ID, "Research"))) {return;}
     html.find('.sub-nav:not(.sub-sub-nav)').append('<a data-tab="subsystems" class="subsystemTab">Subsystems</a>')
-    
+    appID = partySheet.appID
     const targetEl = html[0].querySelector('.container');
     if(targetEl) {
         partySheet.__subsystem_tab = new SubsystemTab({
@@ -71,9 +72,9 @@ Hooks.on('renderPartySheetPF2e', function(partySheet, html, data) {
 	}
 });
 
-Hooks.on('updateActor', function(arg1, arg2, arg3, arg4) {
+Hooks.on('updateActor', function(party, arg2, arg3, arg4) {
 	if(arg2._id==="xxxPF2ExPARTYxxx") {
-		if(arg1._sheet._tabs[0].active==="subsystems") {
+		if(party.apps[appID]._tabs[0].active==="subsystems"){
 			subsystemTabActive = true;
 		} else {
 			subsystemTabActive = false;
@@ -83,11 +84,11 @@ Hooks.on('updateActor', function(arg1, arg2, arg3, arg4) {
 
 function check(target) {
 	if(target.classList.contains("active")) {
-		Helper.log(true, true)
+		//Helper.log(true, true)
 		subsystemTabActive = true;
 		return true;
 	} else {
-		Helper.log(true, false)
+		//Helper.log(true, false)
 		subsystemTabActive = false;
 		return false;
 	}
@@ -97,6 +98,5 @@ Hooks.on('closePartySheetPF2e', function(partySheet, html, data) {
 	partySheet?.__subsystem_tab?.$destroy();
 
 	const targetEl = html[0].querySelector('.subsystemTab')
-	Helper.log(true, targetEl)
 	check(targetEl)
 });