+++ CHANGELOG.md
@@ -1,2 +1,13 @@
+# 0.0.3
+- added Chases subsystem
+- hopefully fixed party sheet tab switching on refresh issue
+
+# 0.0.2
+- fixed issue with build script resulting in broken module, no code changes in this version
+
 # 0.0.1
 initial module setup
+new subsystems included are:
+- Influence
+- Research
+- Victory Points